requirements.txt
google-generativeai
requests
python-dotenv
beautifulsoup4
pyttsx3
SpeechRecognition
pynput
pywin32
